# backend/openapi/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TradeConnect API",
    "version": "0.1.0",
    "description": "TradeConnect backend API contract (v1)."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateJobInput": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "subcategory_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "urgency": {
            "type": "string",
            "enum": [
              "emergency",
              "within_48h",
              "this_week",
              "this_month",
              "flexible"
            ],
            "default": "flexible"
          },
          "property_type": {
            "type": "string",
            "enum": [
              "house",
              "apartment",
              "townhouse",
              "commercial",
              "land",
              "other"
            ]
          },
          "suburb": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "postcode": {
            "type": "string",
            "pattern": "^\\d{4}$"
          },
          "state": {
            "type": "string",
            "enum": [
              "NSW",
              "VIC",
              "QLD",
              "WA",
              "SA",
              "TAS",
              "ACT",
              "NT"
            ]
          },
          "exact_address": {
            "type": "string",
            "minLength": 5,
            "maxLength": 300
          },
          "budget_min": {
            "type": "integer",
            "minimum": 100,
            "maximum": 10000000
          },
          "budget_max": {
            "type": "integer",
            "minimum": 100,
            "maximum": 10000000
          },
          "budget_is_gst": {
            "type": "boolean",
            "default": false
          },
          "preferred_start_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "preferred_end_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "time_window_notes": {
            "type": "string",
            "maxLength": 500
          },
          "publish": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "category_id",
          "title",
          "description",
          "suburb",
          "postcode",
          "state"
        ],
        "additionalProperties": false
      },
...
      "CreateConversationInput": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "job_id",
          "customer_id"
        ],
        "additionalProperties": false
      },
...
      "JobsListResponse": {
        "type": "object",
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "title",
                "status"
              ],
              "additionalProperties": true
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "jobs"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Customer jobs list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/conversations": {
      "get": {
        "tags": [
          "Messaging"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Conversations list"
          }
        }
      },
      "post": {
        "tags": [
          "Messaging"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation opened"
          }
        }
      }
    },
...
  }
}